+++ CHANGES.md
@@ -5,6 +5,9 @@ v1.0.0-alpha.x
 
 - Added `OS.Task.enqueue` to request the task is executed in the next
   run loop.
+- `LatchingButton` now updates on touch-end to allow the control to be
+  left in a unchanged state by ending the touch outside of the control.
+  It now calls on_button_cancel in that situation.
 
 v1.0.0-alpha.3
 ==============